Refactor street splitting code
- rename SimpleStreetSplitter to StreetSplitter as it is the only StreetSplitter in use
- add some checks to make sure duplicate indexing doesn't occur on the same graph
- refactor some areas where duplicate indexing was occurring
