+++ src/main/java/org/opentripplanner/routing/impl/StreetVertexIndexServiceImpl.java
@@ -4,8 +4,6 @@ package org.opentripplanner.routing.impl;
 import org.locationtech.jts.geom.Coordinate;
 import org.locationtech.jts.geom.Envelope;
 import org.locationtech.jts.geom.LineString;
-import org.locationtech.jts.index.SpatialIndex;
-import org.locationtech.jts.index.strtree.STRtree;
 import org.opentripplanner.analyst.core.Sample;
 import org.opentripplanner.analyst.request.SampleFactory;
 import org.opentripplanner.common.geometry.GeometryUtils;
@@ -13,16 +11,19 @@ import org.opentripplanner.common.geometry.HashGridSpatialIndex;
 import org.opentripplanner.common.geometry.SphericalDistanceLibrary;
 import org.opentripplanner.common.model.GenericLocation;
 import org.opentripplanner.common.model.P2;
-import org.opentripplanner.graph_builder.linking.SimpleStreetSplitter;
 import org.opentripplanner.graph_builder.linking.StreetSplitter;
 import org.opentripplanner.routing.core.RoutingRequest;
-import org.opentripplanner.routing.edgetype.*;
+import org.opentripplanner.routing.edgetype.PatternEdge;
+import org.opentripplanner.routing.edgetype.SampleEdge;
+import org.opentripplanner.routing.edgetype.SimpleTransfer;
+import org.opentripplanner.routing.edgetype.StreetEdge;
+import org.opentripplanner.routing.edgetype.TemporaryFreeEdge;
+import org.opentripplanner.routing.edgetype.TemporaryPartialStreetEdge;
 import org.opentripplanner.routing.graph.Edge;
 import org.opentripplanner.routing.graph.Graph;
 import org.opentripplanner.routing.graph.Vertex;
 import org.opentripplanner.routing.location.TemporaryStreetLocation;
 import org.opentripplanner.routing.services.StreetVertexIndexService;
-import org.opentripplanner.routing.util.ElevationUtils;
 import org.opentripplanner.routing.vertextype.SampleVertex;
 import org.opentripplanner.routing.vertextype.StreetVertex;
 import org.opentripplanner.routing.vertextype.TransitStop;
@@ -30,7 +31,10 @@ import org.opentripplanner.util.I18NString;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.Collection;
+import java.util.Iterator;
+import java.util.List;
 
 /**
  * Indexes all edges and transit vertices of the graph spatially. Has a variety of query methods
@@ -47,9 +51,9 @@ public class StreetVertexIndexServiceImpl implements StreetVertexIndexService {
     /**
      * Contains only instances of {@link StreetEdge}
      */
-    private SpatialIndex edgeTree;
-    private SpatialIndex transitStopTree;
-    private SpatialIndex verticesTree;
+    private HashGridSpatialIndex edgeTree;
+    private HashGridSpatialIndex transitStopTree;
+    private HashGridSpatialIndex verticesTree;
 
     // private static final double SEARCH_RADIUS_M = 100; // meters
     // private static final double SEARCH_RADIUS_DEG = DistanceLibrary.metersToDegrees(SEARCH_RADIUS_M);
@@ -70,33 +74,15 @@ public class StreetVertexIndexServiceImpl implements StreetVertexIndexService {
 
     static final Logger LOG = LoggerFactory.getLogger(StreetVertexIndexServiceImpl.class);
 
-    private SimpleStreetSplitter simpleStreetSplitter;
+    private StreetSplitter streetSplitter;
 
     public StreetVertexIndexServiceImpl(Graph graph) {
-        this(graph, true);
-    }
-
-    public StreetVertexIndexServiceImpl(Graph graph, boolean hashGrid) {
         this.graph = graph;
-        if (hashGrid) {
-            edgeTree = new HashGridSpatialIndex<>();
-            transitStopTree = new HashGridSpatialIndex<>();
-            verticesTree = new HashGridSpatialIndex<>();
-        } else {
-            edgeTree = new STRtree();
-            transitStopTree = new STRtree();
-            verticesTree = new STRtree();
-        }
+        edgeTree = new HashGridSpatialIndex<>();
+        transitStopTree = new HashGridSpatialIndex<>();
+        verticesTree = new HashGridSpatialIndex<>();
         postSetup();
-        if (!hashGrid) {
-            ((STRtree) edgeTree).build();
-            ((STRtree) transitStopTree).build();
-            simpleStreetSplitter = new SimpleStreetSplitter(this.graph, null, null);
-        } else {
-            simpleStreetSplitter = new SimpleStreetSplitter(this.graph,
-                    (HashGridSpatialIndex<Edge>) edgeTree, transitStopTree);
-        }
-
+        streetSplitter = new StreetSplitter(graph, edgeTree, transitStopTree);
     }
 
     /**
@@ -212,11 +198,7 @@ public class StreetVertexIndexServiceImpl implements StreetVertexIndexService {
                 if (geometry == null) {
                     continue;
                 }
-                Envelope env = geometry.getEnvelopeInternal();
-                if (edgeTree instanceof HashGridSpatialIndex)
-                    ((HashGridSpatialIndex)edgeTree).insert(geometry, e);
-                else
-                    edgeTree.insert(env, e);
+                edgeTree.insert(geometry, e);
             }
             if (v instanceof TransitStop) {
                 Envelope env = new Envelope(v.getCoordinate());
@@ -317,7 +299,7 @@ public class StreetVertexIndexServiceImpl implements StreetVertexIndexService {
                                        boolean endVertex) {
         Coordinate c = loc.getCoordinate();
         if (c != null) {
-            return simpleStreetSplitter.linkOriginDestination(loc, options, endVertex);
+            return streetSplitter.linkOriginDestination(loc, options, endVertex);
         }
 
         // No Coordinate available.
@@ -370,6 +352,6 @@ public class StreetVertexIndexServiceImpl implements StreetVertexIndexService {
 
     @Override
     public StreetSplitter getStreetSplitter() {
-        return simpleStreetSplitter;
+        return streetSplitter;
     }
 }
