+++ src/main/java/org/opentripplanner/updater/bike_park/BikeParkUpdater.java
@@ -7,7 +7,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
 import java.util.Set;
-import java.util.concurrent.ExecutionException;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import org.opentripplanner.graph_builder.linking.StreetSplitter;
@@ -23,7 +22,7 @@ import org.opentripplanner.updater.JsonConfigurable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import static org.opentripplanner.graph_builder.linking.SimpleStreetSplitter.*;
+import static org.opentripplanner.graph_builder.linking.StreetSplitter.*;
 
 /**
  * Graph updater that dynamically sets availability information on bike parking lots.
